feat(contribution): simplifie la gestion des contributions avec des fonctions réutilisables et améliore la validation des champs requis
